pyproject.toml
[tool.poetry]
name = "vagasayajinbot"
version = "0.1.0"
description = "Vaga Sayajin Bot"
authors = ["Hector"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jobspy = "^0.29.0"
flask = "^3.0.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
